+++ src/main/java/com/z/sharding/util/CRC32Util.java
@@ -4,9 +4,9 @@ import java.util.zip.CRC32;
 
 public class CRC32Util {
 
-    public static int hash(String value) {
+    public static long hash(String value) {
         CRC32 crc = new CRC32();
         crc.update(value.getBytes());
-        return (int)(crc.getValue() & 0xffffffffL);
+        return crc.getValue() & 0xffffffffL;
     }
 }
